Added *basic* support for forwarding emails
Added a new forward.html page along with a new forward view. Emails can now
be forwarded by users by going to an individual email and selecting "Forward".
It essentially populates a variation of the compose form with data from an
already existing email and allows the user to edit it. This creates a new email
and relations with `is_forward` set to True.